[Update]:Created bucketlist endpoints

+++ src/app/api/category/delete-category/[categoryId]/route.ts
@@ -5,13 +5,24 @@ export async function DELETE({ params }: { params: { categoryId: string } }) {
     try {
         const categoryId = params.categoryId;
 
-        const deletedCategory = await prisma.category.delete({
+        const categoryToDelete = await prisma.category.findUnique({
             where: {
                 id: parseInt(categoryId)
             }
         })
 
-        return NextResponse.json({ message: "Category deleted successfully", data: deletedCategory }, { status: 200 })
+        if (categoryToDelete) {
+            await prisma.category.delete({
+                where: {
+                    id: parseInt(categoryId)
+                }
+            })
+
+            return NextResponse.json({ message: "Category deleted successfully" }, { status: 200 })
+        }
+        else {
+            return NextResponse.json({ error: "Category not found" }, { status: 404 })
+        }
     }
     catch (error) {
         return NextResponse.json({ error: "Internal server error" }, { status: 500 })
